Add asset bridging support to fills endpoints
